Use the shared wma run level decoding function

+++ wmapro/wma3data.h
@@ -326,7 +326,7 @@ static const uint8_t coef1_huffbits[HUFF_COEF1_SIZE] = {
 };
 
 
-static const uint8_t coef0_run[HUFF_COEF0_SIZE] = {
+static const uint16_t coef0_run[HUFF_COEF0_SIZE] = {
       0,   0,   0,   1,   2,   3,   4,   5,   6,   7,   8,   9,  10,  11,
      12,  13,  14,  15,  16,  17,  18,  19,  20,  21,  22,  23,  24,  25,
      26,  27,  28,  29,  30,  31,  32,  33,  34,  35,  36,  37,  38,  39,
@@ -349,7 +349,7 @@ static const uint8_t coef0_run[HUFF_COEF0_SIZE] = {
       1,   0,   1,   0,   1,   0,
 };
 
-static const uint8_t coef0_level[HUFF_COEF0_SIZE] = {
+static const uint16_t coef0_level[HUFF_COEF0_SIZE] = {
       0,   0,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,
       1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,
       1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   1,
@@ -373,7 +373,7 @@ static const uint8_t coef0_level[HUFF_COEF0_SIZE] = {
 };
 
 
-static const uint8_t coef1_run[HUFF_COEF1_SIZE] = {
+static const uint16_t coef1_run[HUFF_COEF1_SIZE] = {
      0,  0,  0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 14, 15,
     16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33,
     34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51,
@@ -390,7 +390,7 @@ static const uint8_t coef1_run[HUFF_COEF1_SIZE] = {
      1,  0,  1,  0,  1,  0,  1,  0,  0,  0,
 };
 
-static const uint8_t coef1_level[HUFF_COEF1_SIZE] = {
+static const uint16_t coef1_level[HUFF_COEF1_SIZE] = {
      0,  0,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,
      1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,
      1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,  1,
